splitup vector_renderer
