+++ src/actinia_core/rest/vector_renderer.py
@@ -4,7 +4,7 @@
 # performance processing of geographical data that uses GRASS GIS for
 # computational tasks. For details, see https://actinia.mundialis.de/
 #
-# Copyright (c) 2016-2018 Sören Gebbert and mundialis GmbH & Co. KG
+# Copyright (c) 2016-2022 Sören Gebbert and mundialis GmbH & Co. KG
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -29,20 +29,15 @@ Raster map renderer
 import os
 from flask_restful_swagger_2 import swagger
 from flask import jsonify, make_response, Response
-from tempfile import NamedTemporaryFile
-from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
-     import EphemeralProcessing
 from actinia_core.core.common.redis_interface import enqueue_job
 from actinia_core.rest.base.renderer_base import RendererBaseResource
-from actinia_core.processing.actinia_processing.ephemeral.base.renderer_base \
-     import EphemeralRendererBase
 from actinia_core.models.response_models import ProcessingErrorResponseModel
+from actinia_core.processing.common.vector_renderer import start_job
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert"
-__copyright__ = "Copyright 2016-2018, Sören Gebbert and mundialis GmbH & Co. KG"
-__maintainer__ = "Sören Gebbert"
-__email__ = "soerengebbert@googlemail.com"
+__copyright__ = "Copyright 2016-2022, Sören Gebbert and mundialis GmbH & Co. KG"
+__maintainer__ = "mundialis"
 
 
 class SyncEphemeralVectorRendererResource(RendererBaseResource):
@@ -171,55 +166,3 @@ class SyncEphemeralVectorRendererResource(RendererBaseResource):
                     return Response(image, mimetype='image/png')
 
         return make_response(jsonify(response_model), http_code)
-
-
-def start_job(*args):
-    processing = EphemeralVectorRenderer(*args)
-    processing.run()
-
-
-class EphemeralVectorRenderer(EphemeralRendererBase):
-
-    def __init__(self, *args):
-
-        EphemeralProcessing.__init__(self, *args)
-
-    def _execute(self, skip_permission_check=True):
-        """Render the vector image
-
-        Workflow:
-
-            1. The default region is set to the vector region
-            2. User specific region settings are applied
-            3. d.vect is invoked to create the PNG file
-
-        """
-
-        self._setup()
-
-        vector_name = self.map_name
-        options = self.rdc.user_data
-        self.required_mapsets.append(self.mapset_name)
-
-        with NamedTemporaryFile(suffix=".png") as file:
-            result_file = file.name
-
-        region_pc = self._setup_render_environment_and_region(
-            options=options, result_file=result_file)
-
-        pc = {}
-        pc["1"] = {"module": "g.region", "inputs": {
-            "vector": vector_name + "@" + self.mapset_name}}
-        pc["2"] = region_pc
-        pc["3"] = {
-            "module": "d.vect",
-            "inputs": {"map": vector_name + "@" + self.mapset_name},
-            "flags": "c"}
-
-        # Run the selected modules
-        self.skip_region_check = True
-        process_list = self._create_temporary_grass_environment_and_process_list(
-            process_chain=pc, skip_permission_check=True)
-        self._execute_process_list(process_list)
-
-        self.module_results = result_file
